makefile added

+++ app/build.gradle.kts
@@ -1,4 +1,3 @@
-
 plugins { application }
 
 repositories { mavenCentral() }
@@ -28,12 +27,7 @@ reporting.baseDir = file("reports/gradle")
 tasks.register("cleanReports") {
     doLast {
         val reportsDir = File("$projectDir/reports")
-        if (reportsDir.exists()) {
-            reportsDir.deleteRecursively()
-            println("Deleted $projectDir/reports successfully")
-        }else{
-            println("$projectDir/reports not present")
-        }
+        if (reportsDir.exists()) reportsDir.deleteRecursively()
     }
 }
 
@@ -52,18 +46,10 @@ tasks.named<Test>("test") {
     doLast {
         val buildDir = layout.buildDirectory.dir("allure-results").get().asFile
         buildDir.mkdirs()
-
-        val reportFile = File(buildDir, "environment.properties")
-        val osVersion = System.getProperty("os.version")
-        val osName = System.getProperty("os.name")
-        val javaVersion = System.getProperty("java.version")
-
-        reportFile.writeText("Platform = $osName : $osVersion\nJava = $javaVersion\n")
+        File(buildDir, "environment.properties")
+            .writeText("Platform = "+System.getProperty("os.name")+" : "+System.getProperty("os.version")+"\nJava = "+System.getProperty("java.version"))
     }
 }
 
-tasks.register("allureReport", Exec::class) {
-    commandLine("allure", "generate", "$projectDir/build/allure-results", "--report-dir", "$projectDir/reports/allure",   "--clean")
-}
